requirements.txt
pandas
scikit-learn
xgboost
kafka-python
numpy
scipy
joblib
